06-challenges/29-infinite-scroll
I did not come up with the same solution as provided.
I also misunderstood the direction of the constraint.
I thought of it as `TRow` needed to be an object that had `key` in it.
The relationship being, if `TRow` did not have `key` in it, then that'd
be a type error along the lines of "missing property `<key>`".
According to the solution it's the other way around, where the `key`
must be one of the propties on `TRow`, which changes how you'd defined
the type variables and constraints.

So my initial implementation worked based on the previous assumption, so
I had a type variable `TKey extends string` and then I had `TRow extends
Record<TKey, unknown>`. The type errors would show on the `fetchRows`
function.

The provided solution was a bit simpler. `TRow` was the only type
variable and the `key` had a type definition of `keyof TRow`.
The relationship with this solution bewteen `key` and `TRow`, is one
where if they `key` provided is not one of the properties of `TRow` the
type error would be on the `key` params.

I liked this solution better, less type variables and simpler inference.